refactor(Core/Player): extract KillRewarder, TradeData and PlayerTaxi (#6804)
